--- a/index.html
+++ b/index.html
@@ -4,6 +4,16 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     
+    <!-- Google Analytics -->
+    <script async src="https://www.googletagmanager.com/gtag/js?id=G-L1GL1QWGBR"></script>
+    <script>
+      window.dataLayer = window.dataLayer || [];
+      function gtag(){dataLayer.push(arguments);}
+      gtag('js', new Date());
+
+      gtag('config', 'G-L1GL1QWGBR');
+    </script>
+    
     <!-- Primary SEO Meta Tags -->
     <title>Free Online Clock & Time Zone Converter | EST to IST, PST to IST | DigitalClockApp.com</title>
     <meta name="description" content="Free online clock with time zone converter. Convert EST to IST, PST to IST, CST to IST, UTC to IST. Features: digital clock, timer, stopwatch, Pacific Time, Mountain Time, Central Time, Eastern Time. No download required.">
@@ -37,6 +47,9 @@
     <link rel="shortcut icon" type="image/png" href="./logo.png">
     <link rel="apple-touch-icon" href="./logo.png">
     
+    <!-- Preload Digital7 Font -->
+    <link rel="preload" href="digital-7.ttf" as="font" type="font/ttf" crossorigin>
+    
     <!-- Stylesheet -->
     <link rel="stylesheet" href="styles.css">
     
@@ -69,6 +82,9 @@
     </script>
 </head>
 <body>
+    <!-- Font loading test element -->
+    <div class="font-loading-test">0123456789</div>
+    
     <!-- Header with SEO-friendly content -->
     <header style="display: none;">
         <h1>Free Online Clock & Time Zone Converter</h1>
@@ -191,7 +207,7 @@
                     <span class="timer-separator">:</span>
                     <input type="text" class="timer-input" id="secondsInput" placeholder="00" maxlength="2" inputmode="numeric">
                 </div>
-                <div class="milliseconds-display" id="timerMilliseconds">000ms</div>
+                <div class="milliseconds-display" id="timerMilliseconds">00ms</div>
                 
                 <div class="timer-controls">
                     <div class="action-buttons">
@@ -220,7 +236,7 @@
         <div id="stopwatchMode" class="mode-container">
             <div class="clock-container">
                 <div class="stopwatch-display" id="stopwatchDisplay">00:00:00</div>
-                <div class="milliseconds-display" id="stopwatchMilliseconds">000ms</div>
+                <div class="milliseconds-display" id="stopwatchMilliseconds">00ms</div>
                 
                 <div class="stopwatch-controls">
                     <div class="action-buttons">